feat: Add GPXFileButton and FormField components
- Added GPXFileButton component for opening GPX files.
- Added FormField component for rendering form fields with labels and initial values.
