+++ assets/storage/tasks/expected-output/extracted-task-file-sample/Attachments/a81e8752-7e9d-468f-a254-180028a09c0b.gpx
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="utf-8"?>
+<gpx version="1.1" creator="PowerForm">
+  <metadata>
+    <farmer>
+      <name>DELA TORRE, RAMIL O</name>
+    </farmer>
+    <time>2024-04-08T13:31:00.3969773+08:00</time>
+  </metadata>
+  <trk>
+    <name>DELA TORRE, RAMIL O</name>
+    <trkseg>
+      <trkpt lat="14.6531133" lon="121.0351767">
+        <time>2024-04-08T13:29:02+08:00</time>
+      </trkpt>
+      <trkpt lat="14.6531133" lon="121.0351767">
+        <time>2024-04-08T13:29:05+08:00</time>
+      </trkpt>
+      <trkpt lat="14.6531133" lon="121.0351767">
+        <time>2024-04-08T13:29:08+08:00</time>
+      </trkpt>
+      <trkpt lat="14.6531133" lon="121.0351767">
+        <time>2024-04-08T13:29:11+08:00</time>
+      </trkpt>
+      <trkpt lat="14.6531133" lon="121.0351767">
+        <time>2024-04-08T13:29:14+08:00</time>
+      </trkpt>
+      <trkpt lat="14.6531133" lon="121.0351767">
+        <time>2024-04-08T13:29:17+08:00</time>
+      </trkpt>
+      <trkpt lat="14.6531133" lon="121.0351767">
+        <time>2024-04-08T13:29:20+08:00</time>
+      </trkpt>
+      <trkpt lat="14.6531133" lon="121.0351767">
+        <time>2024-04-08T13:29:22+08:00</time>
+      </trkpt>
+      <trkpt lat="14.6531133" lon="121.0351767">
+        <time>2024-04-08T13:29:02+08:00</time>
+      </trkpt>
+    </trkseg>
+  </trk>
+</gpx>
